NO-REF: Refactor catalog process
